Create hugginface_ner_classifier.py
This script is unfinished. It operates on the post format for S2Underground's Telegram channel (https://t.me/s2_underground_project). It uses two language models from HuggingFace: one is an NER model to recgnize and tag certain information contained within the channel's posts (LOC, PER, ORG) and the other is an event classifier that classifies events in each post (each feature is a unique location) with a keyword based on a confidence score. This logic is very similar to the confidence scores determined in the script classifying CSV features layers based on their relationship to various PUG topics.